Improved osgi tests and simplified code
