+++ tests/dozer-osgi-tests/src/test/java/com/github/dozermapper/osgitests/AbstractDozerCoreOsgiContainerTest.java
@@ -15,75 +15,29 @@
  */
 package com.github.dozermapper.osgitests;
 
-import javax.inject.Inject;
+import java.util.Dictionary;
+import java.util.List;
 
 import com.github.dozermapper.core.DozerBeanMapperBuilder;
 import com.github.dozermapper.core.Mapper;
 import com.github.dozermapper.core.osgi.Activator;
 import com.github.dozermapper.core.osgi.OSGiClassLoader;
-import com.github.dozermapper.osgitests.karaf.BundleOptions;
-import com.github.dozermapper.osgitests.support.OsgiTestSupport;
+import com.github.dozermapper.osgitests.support.PaxExamTestSupport;
 import com.github.dozermapper.osgitestsmodel.Person;
 
-import org.junit.After;
-import org.junit.Before;
+import org.apache.felix.utils.manifest.Clause;
+import org.apache.felix.utils.manifest.Parser;
 import org.junit.Test;
-import org.ops4j.pax.exam.Configuration;
-import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.ProbeBuilder;
-import org.ops4j.pax.exam.TestProbeBuilder;
 import org.osgi.framework.Bundle;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.Constants;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.osgi.framework.wiring.BundleCapability;
+import org.osgi.framework.wiring.BundleRevision;
+import org.osgi.framework.wiring.BundleWiring;
 
+import static junit.framework.TestCase.assertTrue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.ops4j.pax.exam.CoreOptions.junitBundles;
-import static org.ops4j.pax.exam.CoreOptions.options;
 
-public abstract class AbstractDozerCoreOsgiContainerTest extends OsgiTestSupport {
-
-    protected static final Logger LOG = LoggerFactory.getLogger(AbstractDozerCoreOsgiContainerTest.class);
-
-    @Inject
-    protected BundleContext bundleContext;
-
-    @ProbeBuilder
-    public TestProbeBuilder probeConfiguration(TestProbeBuilder probe) {
-        // makes sure the generated Test-Bundle contains this import!
-        probe.setHeader(Constants.DYNAMICIMPORT_PACKAGE, "*");
-        return probe;
-    }
-
-    @Before
-    public void setUp() {
-        LOG.info("setUp() using BundleContext: {}", bundleContext);
-    }
-
-    @After
-    public void tearDown() {
-        LOG.info("tearDown()");
-    }
-
-    @Configuration
-    public Option[] config() {
-        return options(
-                // Framework
-                containerConfigOptions(),
-                // Bundles
-                optionalBundles(),
-                BundleOptions.coreBundles(),
-                junitBundles()
-        );
-    }
-
-    protected abstract Option containerConfigOptions();
-
-    protected Option optionalBundles() {
-        return BundleOptions.optionalBundles();
-    }
+public abstract class AbstractDozerCoreOsgiContainerTest extends PaxExamTestSupport {
 
     @Test
     public void canGetBundleFromDozerCore() {
@@ -96,6 +50,8 @@ public abstract class AbstractDozerCoreOsgiContainerTest extends OsgiTestSupport
         assertEquals(Bundle.ACTIVE, core.getState());
 
         for (Bundle current : bundleContext.getBundles()) {
+            LOG.info("Bundle: {}", current.getSymbolicName());
+
             //Ignore any Karaf bundles
             if (current.getSymbolicName().startsWith("org.apache.karaf")
                 || current.getSymbolicName().startsWith("org.jline")) {
@@ -114,6 +70,7 @@ public abstract class AbstractDozerCoreOsgiContainerTest extends OsgiTestSupport
                 .build();
 
         assertNotNull(mapper);
+        assertNotNull(mapper.getMappingMetadata());
     }
 
     @Test
@@ -123,10 +80,53 @@ public abstract class AbstractDozerCoreOsgiContainerTest extends OsgiTestSupport
                 .withClassLoader(new OSGiClassLoader(com.github.dozermapper.osgitestsmodel.Activator.getBundleContext()))
                 .build();
 
+        assertNotNull(mapper);
+        assertNotNull(mapper.getMappingMetadata());
+
         Person answer = mapper.map(new Person("bob"), Person.class);
 
         assertNotNull(answer);
         assertNotNull(answer.getName());
         assertEquals("bob", answer.getName());
+    }
+
+    @Test
+    public void canResolveAllImports() {
+        Bundle core = getBundle(bundleContext, "com.github.dozermapper.dozer-core");
+        assertNotNull(core);
+        assertEquals(Bundle.ACTIVE, core.getState());
+
+        Dictionary<String, String> headers = core.getHeaders();
+        assertNotNull(headers);
+
+        Clause[] clauses = Parser.parseHeader(headers.get("Import-Package"));
+        for (Clause clause : clauses) {
+            LOG.info("Package: {}", clause.getName());
+
+            assertTrue(checkPackage(clause.getName()));
+        }
+    }
+
+    private boolean checkPackage(String packageName) {
+        Bundle[] bundles = bundleContext.getBundles();
+        for (int i = 0; (bundles != null) && (i < bundles.length); i++) {
+            BundleWiring wiring = bundles[i].adapt(BundleWiring.class);
+            List<BundleCapability> caps = wiring == null ? null : wiring.getCapabilities(BundleRevision.PACKAGE_NAMESPACE);
+            if (caps != null) {
+                for (BundleCapability cap : caps) {
+                    String n = getAttribute(cap, BundleRevision.PACKAGE_NAMESPACE);
+                    if (packageName.equals(n)) {
+                        return true;
+                    }
+                }
+            }
+        }
+
+        return false;
+    }
+
+    private String getAttribute(BundleCapability cap, String name)  {
+        Object obj = cap.getAttributes().get(name);
+        return obj != null ? obj.toString() : null;
     }
 }
