+++ tests/dozer-osgi-tests/src/test/java/com/github/dozermapper/osgitests/CoreKarafContainerTest.java
@@ -22,12 +22,37 @@ import com.github.dozermapper.core.osgi.OSGiClassLoader;
 import com.github.dozermapper.osgitestsmodel.Person;
 
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.ops4j.pax.exam.Configuration;
+import org.ops4j.pax.exam.Option;
+import org.ops4j.pax.exam.junit.PaxExam;
+import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
+import org.ops4j.pax.exam.spi.reactors.PerClass;
 
+import static com.github.dozermapper.osgitests.support.BundleOptions.coreBundles;
+import static com.github.dozermapper.osgitests.support.BundleOptions.optionalBundles;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.ops4j.pax.exam.CoreOptions.junitBundles;
+import static org.ops4j.pax.exam.CoreOptions.options;
 
-public abstract class DozerCoreOsgiContainerTest extends AbstractDozerCoreOsgiContainerTest {
+@RunWith(PaxExam.class)
+@ExamReactorStrategy(PerClass.class)
+public class CoreKarafContainerTest extends AbstractDozerCoreOsgiContainerTest {
+
+    @Configuration
+    public Option[] config() {
+        return options(
+                // Framework
+                karaf4ContainerConfigOptions(),
+
+                // Bundles
+                optionalBundles(),
+                coreBundles(),
+                junitBundles()
+        );
+    }
 
     @Test
     public void canMapUsingXMLWithVariables() {
