--- tests/dozer-osgi-tests/src/test/java/com/github/dozermapper/osgitests/karaf/DozerCoreKaraf4OsgiContainerTest.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2005-2018 Dozer Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.github.dozermapper.osgitests.karaf;
-
-import com.github.dozermapper.osgitests.DozerCoreOsgiContainerTest;
-
-import org.junit.runner.RunWith;
-import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.junit.PaxExam;
-import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
-import org.ops4j.pax.exam.spi.reactors.PerClass;
-
-@RunWith(PaxExam.class)
-@ExamReactorStrategy(PerClass.class)
-public class DozerCoreKaraf4OsgiContainerTest extends DozerCoreOsgiContainerTest {
-
-    @Override
-    protected Option containerConfigOptions() {
-        return KarafOptions.karaf4ContainerConfigOptions();
-    }
-}
